Attempted manual merge from cnlohr's master. May or may not have worked

+++ colorchord2/DisplayOUTDriver.c
@@ -88,7 +88,7 @@ static void DPOUpdate(void * id, struct NoteFinder*nf)
 	for( i = 0; i < d->note_peaks; i++ )
 	{
 		struct LINote * l = &d->notes[i];
-		l->ledexp = pow( nf->note_amplitudes2[i], d->pow ) - d->cutoff;
+		l->ledexp = powf( nf->note_amplitudes2[i], d->pow ) - d->cutoff;
 		if( l->ledexp < 0 ) l->ledexp = 0;
 		totalexp += l->ledexp;
 	}
